+++ tests/conftest.py
@@ -63,6 +63,12 @@ def coastal_swn(coastal_lines_gdf):
     return swn.SurfaceWaterNetwork.from_lines(coastal_lines_gdf.geometry)
 
 
+@pytest.fixture(scope="module")
+def coastal_swn_w_poly(coastal_lines_gdf, coastal_polygons_gdf):
+    return swn.SurfaceWaterNetwork.from_lines(
+        coastal_lines_gdf.geometry, coastal_polygons_gdf.geometry)
+
+
 @pytest.fixture(scope="session", autouse=True)
 def coastal_flow_ts():
     csv_fname = "streamq_20170115_20170128_topnet_03046727_m3day.csv"